Transform editor in a single component
